pyproject.toml
[project]
name = "technitium-api-proxy"
version = "0.1.0"
description = "Lightweight reverse proxy for Technitium DNS Server HTTP API with fine-grained access policies"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "pydantic>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["proxy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "dnspython>=2.8.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests requiring a running Technitium DNS server",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
